backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg2-binary>=2.9
python-dotenv>=1.0

